+++ Source/Headers/objects.h
@@ -108,8 +108,8 @@ void CreateCollisionBoxFromBoundingBox(ObjNode *theNode, float tweakXZ, float tw
 void CreateCollisionBoxFromBoundingBox_Maximized(ObjNode *theNode);
 void CreateCollisionBoxFromBoundingBox_Rotated(ObjNode *theNode, float tweakXZ, float tweakY);
 void CreateCollisionBoxFromBoundingBox_Update(ObjNode *theNode, float tweakXZ, float tweakY);
-ObjNode *FindClosestCType(OGLPoint3D *pt, u_long ctype, Boolean notThruSolid);
-ObjNode *FindClosestCType3D(OGLPoint3D *pt, u_long ctype);
+ObjNode *FindClosestCType(OGLPoint3D *pt, uint32_t ctype, Boolean notThruSolid);
+ObjNode *FindClosestCType3D(OGLPoint3D *pt, uint32_t ctype);
 extern	void KeepOldCollisionBoxes(ObjNode *theNode);
 void AddCollisionBoxToObject(ObjNode *theNode, float top, float bottom, float left,
 							 float right, float front, float back);
