+++ Source/Headers/sprites.h
@@ -10,11 +10,12 @@ enum
 };
 
 
+// Read in from file
 typedef struct
 {
-	long			width,height;
+	int32_t			width,height;
 	float			aspectRatio;			// h/w
-	long			srcFormat, destFormat;
+	int32_t			srcFormat, destFormat;
 	MetaObjectPtr	materialObject;
 }SpriteType;
 
@@ -26,7 +27,7 @@ void LoadSpriteFile(FSSpec *spec, int groupNum, OGLSetupOutputType *setupInfo);
 ObjNode *MakeSpriteObject(NewObjectDefinitionType *newObjDef, OGLSetupOutputType *setupInfo);
 void BlendAllSpritesInGroup(short group);
 void ModifySpriteObjectFrame(ObjNode *theNode, short type, OGLSetupOutputType *setupInfo);
-void DrawSprite(int	group, int type, float x, float y, float scale, float rot, u_long flags, const OGLSetupOutputType *setupInfo);
+void DrawSprite(int	group, int type, float x, float y, float scale, float rot, uint32_t flags, const OGLSetupOutputType *setupInfo);
 void BlendASprite(int group, int type);
 
 ObjNode *MakeFontStringObject(const Str31 s, NewObjectDefinitionType *newObjDef, OGLSetupOutputType *setupInfo, Boolean center);
