+++ Source/Screens/HighScores.c
@@ -17,7 +17,7 @@ extern	short	gPrefsFolderVRefNum;
 extern	long	gPrefsFolderDirID;
 extern	FSSpec	gDataSpec;
 extern	OGLSetupOutputType		*gGameViewInfoPtr;
-extern	u_long			gScore,gGlobalMaterialFlags,gLoadedScore;
+extern	uint32_t			gScore,gGlobalMaterialFlags,gLoadedScore;
 extern	Boolean			gPlayingFromSavedGame,gAllowAudioKeys;
 extern	AGLContext		gAGLContext;
 extern	OGLColorRGB			gGlobalColorFilter;
@@ -33,8 +33,8 @@ static void DrawScoreVerbage(OGLSetupOutputType *info);
 static void DrawHighScoresAndCursor(OGLSetupOutputType *info);
 static void SetHighScoresSpriteState(void);
 static void StartEnterName(void);
-static Boolean IsThisScoreInList(u_long score);
-static short AddNewScore(u_long newScore);
+static Boolean IsThisScoreInList(uint32_t score);
+static short AddNewScore(uint32_t newScore);
 static void SaveHighScores(void);
 static void MoveScoreCyc(ObjNode *theNode);
 
@@ -617,7 +617,7 @@ char				blank[MAX_NAME_LENGTH] = "               ";
 // Returns high score slot that score was inserted to or -1 if none
 //
 
-static short AddNewScore(u_long newScore)
+static short AddNewScore(uint32_t newScore)
 {
 short	slot,i;
 
@@ -651,7 +651,7 @@ got_slot:
 // Returns True if this score value is anywhere in the high scores already
 //
 
-static Boolean IsThisScoreInList(u_long score)
+static Boolean IsThisScoreInList(uint32_t score)
 {
 short	slot;
