+++ Source/System/Collision.c
@@ -54,10 +54,10 @@ Boolean			gSolidTriggerKeepDelta;
 // INPUT: startNumCollisions = value to start gNumCollisions at should we need to keep existing data in collision list
 //
 
-void CollisionDetect(ObjNode *baseNode, u_long CType, short startNumCollisions)
+void CollisionDetect(ObjNode *baseNode, uint32_t CType, short startNumCollisions)
 {
 ObjNode 	*thisNode;
-u_long		sideBits,cBits,cType;
+uint32_t		sideBits,cBits,cType;
 short		numBaseBoxes,targetNumBoxes,target;
 CollisionBoxType *baseBoxList;
 CollisionBoxType *targetBoxList;
@@ -301,7 +301,7 @@ next:
 // OUTPUT: totalSides
 //
 
-Byte HandleCollisions(ObjNode *theNode, u_long cType, float deltaBounce)
+Byte HandleCollisions(ObjNode *theNode, uint32_t cType, float deltaBounce)
 {
 Byte		totalSides;
 short		i;
@@ -390,7 +390,7 @@ again:
 		{
 				/* SEE IF THIS OBJECT HAS SINCE BECOME INVALID */
 
-			u_long	targetCType = targetObj->CType;						// get ctype of hit obj
+			uint32_t	targetCType = targetObj->CType;						// get ctype of hit obj
 			if (targetCType == INVALID_NODE_FLAG)
 				continue;
 
@@ -994,7 +994,7 @@ signed char	wind;										// current winding number
 // OUTPUT: # collisions detected
 //
 
-short DoSimplePointCollision(OGLPoint3D *thePoint, u_long cType, ObjNode *except)
+short DoSimplePointCollision(OGLPoint3D *thePoint, uint32_t cType, ObjNode *except)
 {
 ObjNode	*thisNode;
 short	targetNumBoxes,target;
@@ -1079,7 +1079,7 @@ next:
 //
 
 short DoSimpleBoxCollision(float top, float bottom, float left, float right,
-						float front, float back, u_long cType)
+						float front, float back, uint32_t cType)
 {
 ObjNode			*thisNode;
 short			targetNumBoxes,target;
@@ -1317,7 +1317,7 @@ CollisionBoxType *targetBoxList;
 // box here.
 //
 
-float FindHighestCollisionAtXZ(float x, float z, u_long cType)
+float FindHighestCollisionAtXZ(float x, float z, uint32_t cType)
 {
 ObjNode	*thisNode;
 short	targetNumBoxes,target;
@@ -1415,7 +1415,7 @@ next:
 
 /******************** SEE IF LINE SEGMENT HITS ANYTHING **************************/
 
-Boolean SeeIfLineSegmentHitsAnything(const OGLPoint3D *endPoint1, const OGLPoint3D *endPoint2, const ObjNode *except, u_long ctype)
+Boolean SeeIfLineSegmentHitsAnything(const OGLPoint3D *endPoint1, const OGLPoint3D *endPoint2, const ObjNode *except, uint32_t ctype)
 {
 ObjNode	*thisNode;
 OGLPoint2D	p1,p2,crossBeamP1,crossBeamP2;
