+++ Source/System/Objects.c
@@ -19,7 +19,7 @@ extern	float					gFramesPerSecondFrac,gGlobalTransparency;
 extern	OGLSetupOutputType		*gGameViewInfoPtr;
 extern	MOVertexArrayData		**gLocalTriMeshesOfSkelType;
 extern	OGLColorRGB				gGlobalColorFilter;
-extern	u_long					gGlobalMaterialFlags;
+extern	uint32_t					gGlobalMaterialFlags;
 extern	OGLBoundingBox			gObjectGroupBBoxList[MAX_BG3D_GROUPS][MAX_OBJECTS_IN_GROUP];
 
 /****************************/
@@ -1475,7 +1475,7 @@ void UpdateObjectTransforms(ObjNode *theNode)
 {
 OGLMatrix4x4	m,m2;
 OGLMatrix4x4	mx,my,mz,mxz;
-u_long			bits;
+uint32_t			bits;
 
 	if (theNode->CType == INVALID_NODE_FLAG)		// see if already deleted
 		return;
@@ -1584,7 +1584,7 @@ MOMatrixObject	*mo = theNode->BaseTransformObject;
 
 /********************* FIND CLOSEST CTYPE *****************************/
 
-ObjNode *FindClosestCType(OGLPoint3D *pt, u_long ctype, Boolean notThruSolid)
+ObjNode *FindClosestCType(OGLPoint3D *pt, uint32_t ctype, Boolean notThruSolid)
 {
 ObjNode		*thisNodePtr,*best = nil;
 float	d,minDist = 10000000;
@@ -1626,7 +1626,7 @@ next:
 
 /********************* FIND CLOSEST CTYPE 3D *****************************/
 
-ObjNode *FindClosestCType3D(OGLPoint3D *pt, u_long ctype)
+ObjNode *FindClosestCType3D(OGLPoint3D *pt, uint32_t ctype)
 {
 ObjNode		*thisNodePtr,*best = nil;
 float	d,minDist = 10000000;
