+++ Source/System/input.c
@@ -313,7 +313,7 @@ long					bestValue, value, elementType;
 	{
 		for (n = 0; n < NUM_CONTROL_NEEDS; n++)
 		{
-			u_long value;
+			uint32_t value;
 
 			gControlNeeds[n].oldValue = gControlNeeds[n].value;				// remember what the old value was before we update it
